requirements.txt
requests
PyYAML
pytest
PyNaCl
